feat: allow search field in addField
