+++ packages/rath-client/public/locales/en-US.json
@@ -39,6 +39,11 @@
             "qt_25": "25th Percentile",
             "qt_75": "75th Percentile",
             "qt_50": "Median"
+        },
+        "search": {
+            "searchViews": "Search Views",
+            "searchFields": "Search fields / variables",
+            "notFound": "No results found"
         }
     },
     "menu": {
