+++ packages/rath-client/public/locales/zh-CN.json
@@ -39,6 +39,11 @@
             "qt_25": "25%分位数",
             "qt_75": "75%分位数",
             "qt_50": "中位数"
+        },
+        "search": {
+            "searchViews": "搜索视图",
+            "searchFields": "搜索字段/变量",
+            "notFound": "未找到匹配结果"
         }
     },
     "menu": {
