+++ packages/rath-client/src/pages/semiAutomation/semiEmbed/liteFocusZone.tsx
@@ -5,9 +5,9 @@ import intl from 'react-intl-universal';
 import { IFieldMeta } from '../../../interfaces';
 import { useGlobalStore } from '../../../store';
 import ViewField from '../../megaAutomation/vizOperation/viewField';
-import FieldPlaceholder from '../../../components/fieldPlaceholder';
+import FieldPlaceholder from '../../../components/fieldPill/fieldPlaceholder';
 import { MainViewContainer } from '../components';
-import FilterCreationPill from '../../../components/filterCreationPill';
+import FilterCreationPill from '../../../components/fieldPill/filterCreationPill';
 import Narrative from '../narrative';
 import MainCanvas from '../focusZone/mainCanvas';
 import MiniFloatCanvas from '../focusZone/miniFloatCanvas';
